feat: Add functionality to the dentist selection screen.
