+++ frontend/src/sections/ShiftManager/Modal/EditShift.jsx
@@ -39,7 +39,7 @@ export default function EditShift({
   });
 
   const SHIFT_ID = Number(eventInfo.id);
-  //console.log(eventInfo);
+  console.log("event info por props", eventInfo);
   useEffect(() => {
     if (eventInfo.extendedProps) {
       //setea la fecha
@@ -54,12 +54,12 @@ export default function EditShift({
       //setea la hora
 
       const parsedHour = parse(
-        eventInfo.extendedProps.hour,
+        eventInfo.extendedProps.time,
         "HH:mm",
         new Date()
       );
       setSelectedHour(parsedHour);
-      setValue("hour", eventInfo.extendedProps.hour);
+      setValue("hour", eventInfo.extendedProps.time);
       //setea el dentista
       setValue("odontologist", eventInfo.extendedProps.dentistId);
     }
@@ -96,13 +96,15 @@ export default function EditShift({
 
         const updatedEvent = {
           id: SHIFT_ID,
-          title: `Paciente: ${formData.patient_id}`, // Ajusta esto según la estructura de tus datos
+          title: `${formData.title}`, // Ajusta esto según la estructura de tus datos
           start: `${formData.date}T${formData.time}`,
+          backgroundColor: eventInfo.backgroundColor,
+          borderColor: eventInfo.borderColor,
           extendedProps: {
-            ...formData,
+            ...eventInfo.extendedProps,
           },
         };
-
+        console.log("UPDATE EVENTS", updatedEvent);
         updateEventInState(updatedEvent);
         setModalModifyIsVisible(false);
       }
@@ -213,7 +215,7 @@ export default function EditShift({
                     <Controller
                       control={control}
                       name="hour"
-                      defaultValue={eventInfo.extendedProps.hour}
+                      defaultValue={selectedHour}
                       render={({ field }) => (
                         <DatePicker
                           className={`bg-[#F6FBFF] rounded-[4px] w-full border placeholder:text-[#1C3454] placeholder:text-opacity-25 placeholder:text-lg placeholder:font-normal border-[#193B67] border-opacity-15`}
@@ -232,7 +234,7 @@ export default function EditShift({
                           icon={
                             <FiClock className="text-[#1B2B41] text-opacity-70 absolute right-0 pointer-events-none top-1/2 transform -translate-y-1/2 text-2xl" />
                           }
-                          placeholderText={eventInfo.extendedProps.hour}
+                          placeholderText={eventInfo.extendedProps.time}
                           onChange={(hour) => {
                             handleHourChange(hour);
                             field.onChange(format(hour, "HH:mm")); // para cambie el valor del input
